Drop quotes when it confuses more than anything else
This is a not an assignement. Here we are looking for a specific value
in a file and the value that we are looking for does not include the
quotes (this works because the shell strips them but they are really
not needed, it's a bare string without spaces).

Gbp-Dch: Ignore