Drop default value for --cache-dir parameter
It is not at all appropriate for debootstrap to start putting files
in APT's cache of .deb files (it could possibly use it in a read-only
way, but writing to it is not desirable). Furthermore the code was not
working as expected, the default value was only put in place if the
--cache-dir option was passed and it would require an explicit value
anyway.

Also fix the manual page to have the proper spelling of the parameter.