Keep all script files in the udeb
Among the scripts that were dropped in
8ab54a911b27b8eda630740c3c0c45aecfd2feca are the scripts of Debian
derivatives. This means that derivatives would have to rebuild
the package to have a working installer. This is rather unfriendly
for derivatives that make the effort to have their script merged
in the official debootstrap package.

The few kilobytes saved are not worth the trouble, thus we simply
keep all the script files that are provided.

This change thus restores the possibility for derivatives to use the udeb
built by Debian.

Related to #893713.

Acked-by: Cyril Brulebois <kibi@debian.org>