Revert the addition of (mandatory) security support (Closes: #953759).
debootstrap's goal is to provide with a basic Debian system that many
provisioning systems are built upon. Changing behaviour now seems to be
doing a disservice to all those tools that would need adapting. In
addition to not seeming to be a good idea, that triggers regressions for
a number of use cases (also-Closes: #953588, #953593, #953594, #953617).