extract_release_components: Avoid duplicate components in `sources.list`
Currently, adding extra suites, duplicate components are added to
`/etc/apt/sources.list`.

    $ sudo debootstrap --arch=amd64 --variant=minbase --extra-suites=bullseye-proposed-updates bullseye deboottest
    […]
    $ less deboottest/etc/apt/sources.list
    deb https://deb.debian.org/debian bullseye main main
    deb https://deb.debian.org/debian bullseye-proposed-updates main main main main

The reason is the introduction of extra suites, introducing a loop in
`download_release_indices()`, and, therefore,
`extract_release_components()` can be called several times, once for
each suite. Currently, it appends the components unconditionally. So
ensure, components are only appended to the string, if it’s not present
yet, making it unique.

    deb https://deb.debian.org/debian bullseye main
    deb https://deb.debian.org/debian bullseye-proposed-updates main main

(This is still not correct, and is fixed in the next commit.)

Link: https://bugs.debian.org/690210#94
Fixes: 52ff0e7f ("Undo changelog note")