setup_apt_sources: Reset variable in loop to avoid duplicate components
Currently for each suite, the components are duplicated in
`/etc/apt/sources.list` in the next line.

    $ sudo debootstrap --arch=amd64 --variant=minbase --extra-suites=bullseye-proposed-updates bullseye deboottest
    […]
    $ more deboottest/etc/apt/sources.list
    deb https://deb.debian.org/debian bullseye main
    deb https://deb.debian.org/debian bullseye-proposed-updates main main

The extra suites implementation, added in commit 52ff0e7f (Undo
changelog note), adds a for loop storing the components in the variable
`cs`. Not resetting the variable causes the string to be expanded each
time, causing the issue. So, reset it to the empty string at each
iteration, resulting in the wanted behavior.

    deb https://deb.debian.org/debian bullseye main
    deb https://deb.debian.org/debian bullseye-proposed-updates main main

Solution-by: JH Chatenet <jhcha54008@free.fr>
Link: https://bugs.debian.org/690210#94
Fixes: 52ff0e7f ("Undo changelog note")